feat(nfts): shared media component for images and video
