+++ src/analytics/Events.tsx
@@ -613,8 +613,7 @@ export enum AssetsEvents {
 
 export enum NftEvents {
   nft_error_screen_open = 'nft_error_screen_open', // When the high level error screen is mounted
-  nft_image_load = 'nft_image_load', // When of attempting to load an NFT image
-  nft_video_load = 'nft_video_load', // When of attempting to load an NFT video
+  nft_media_load = 'nft_media_load', // When of attempting to load NFT media
   nft_gallery_screen_open = 'nft_gallery_screen_open', // When the gallery screen is mounted
 }
